+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 * Link to https://danger.systems/ in Bitbucket Server comments - HeEAaD
+* Add `markdown_link` for the Bitbucket Server plugin - HeEAaD
 
 ## 3.2.1